+++ src/briefy/leica/models/workflows/project.py
@@ -1,0 +1,17 @@
+"""Project Workflow."""
+from briefy.common.workflow import BriefyWorkflow
+from briefy.common.workflow import WorkflowState
+
+import logging
+
+
+logger = logging.getLogger(__name__)
+
+
+class ProjectWorkflow(BriefyWorkflow):
+    """Workflow for a Project."""
+
+    entity = 'projects'
+    initial_state = 'created'
+
+    created = WorkflowState('created', title='Created', description='Project created')