+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
  * Change role to group in the Asset and Job workflows. (rudaporto)
  * Integrate workflow fix in briefy.common. (rudaporto)
  * BODY-62: Implement pagination. (ericof)
+ * LEICA-63: Improve workflows. (ericof)
 
 1.1.0 (2016-10-04)
 ------------------